+++ codex-rs/core/tests/suite/mod.rs
@@ -10,7 +10,6 @@ mod cli_stream;
 mod client;
 mod compact;
 mod compact_resume_fork;
-mod error_truncation;
 mod exec;
 mod fork_conversation;
 mod grep_files;
@@ -32,9 +31,9 @@ mod shell_serialization;
 mod stream_error_allows_next_turn;
 mod stream_no_completed;
 mod tool_harness;
-mod tool_output_truncation;
 mod tool_parallelism;
 mod tools;
+mod truncation;
 mod unified_exec;
 mod user_notification;
 mod view_image;
